+++ src/lib/traversal.ts
@@ -23,12 +23,14 @@ export function mapErrors<T extends AnyZodObject>(
   const output: Record<string, unknown> = {};
   const entries = Object.entries(obj);
 
-  if (
+  if (top && '_errors' in obj) {
+    if (obj._errors.length) output._errors = obj._errors;
+  } else if (
     entries.length === 1 &&
     entries[0][0] === '_errors' &&
     obj._errors.length
   ) {
-    return (top ? obj : obj._errors) as ValidationErrors<T>;
+    return obj._errors as unknown as ValidationErrors<T>;
   } else if (obj._errors.length) {
     output._errors = obj._errors;
   }